pyproject.toml
[project]
name = "sensor-fault-env"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi",
    "uvicorn",
    "openai",
    "requests",
]

[project.scripts]
server = "server.app:main"